backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
"rest_framework_simplejwt.token_blacklist"
django-environ
django-filter

google-auth 
google-auth-oauthlib 
google-auth-httplib2
django-cors-headers
